% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bipartiteMatch}
\alias{bipartiteMatch}
\title{Bipartite matching with the Hungarian algorithm}
\usage{
bipartiteMatch(costMatrix)
}
\arguments{
\item{costMatrix}{a cost matrix on which to match, note that the best matches will minimize the diagonal, thus distances are appropriate.}
}
\value{
List with cost and pairings, pairings are given as an nx2 matrix giving edges that are metched.
}
\description{
Solves weighted bipartite matching on a cost matrix (i.e. distance) using the hungarian algorithm
}
\details{
This is a copy/clone of the code developed by Cong Ma and the Rcpp wrapper developed by Justin Silverman released in the RcppHungarian package.
}
