% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{rank}
\alias{rank}
\title{Rank features by importance}
\usage{
rank(h, d = NULL)
}
\arguments{
\item{h}{sample factor model}

\item{d}{NULL, or the existing diagonal to update}
}
\value{
list of "d" and "h". "d" encodes the scaling diagonal vector (rank of each feature), and "h" is the newly scaled sample embeddings matrix.
}
\description{
Finds a diagonal vector \code{d} such that rows in \code{h} sum to 1 when multiplied by \code{d}. In other words, the diagonal vector gives the relative weights of each factor across all samples in the model and a scaling of \code{h} which aids interpretability.
}
\details{
The scaling is applied to the short edge of h. This assumes that the number of samples/features is greater than the rank.
}
