% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_bls}
\alias{Rcpp_bls}
\title{Bounded least squares}
\usage{
Rcpp_bls(x, a, b, maxit, tol, min, max, fixed)
}
\arguments{
\item{x}{required initial numeric vector of doubles for x, may simply specify a vector of ones}

\item{a}{square numeric dense matrix of doubles containing the coefficients of the linear system}

\item{b}{numeric vector of doubles giving the right-hand side of the linear system}

\item{maxit}{integer, maximum number of iterations for least squares sequential coordinate descent}

\item{tol}{double, tolerance for convergence of the least squares sequential coordinate descent solver}

\item{min}{double minimum permitted value, use zero for NMF}

\item{max}{double, maximum permitted value, use a large value to avoid an upper bound constraint}

\item{fixed}{a vector of 0 or 1 corresponding to "x" indicating if values in the supplied "x" vector should be fixed (i.e. not updated). To fix, specify 1, otherwise specify 0. If not fixing any values, specify a vector of zeros of length x.}
}
\value{
x vector of least squares solution
}
\description{
Solve \code{a \%*\% x = b} using sequential coordinate descent bounded least squares, where \code{a} is a square matrix and \code{x} and \code{b} are vectors.
}
\details{
See \code{cls} R function documentation for a more convenient interface and details. This function should only be used when speed is paramount. Fields are not nullable.
}
