% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_L0_bls}
\alias{Rcpp_L0_bls}
\title{L0 and/or multinomial-constrained bounded least squares}
\usage{
Rcpp_L0_bls(x, a, b, maxit, tol, min, max, values, L0, fixed)
}
\arguments{
\item{x}{required initial numeric vector of doubles for x, may simply specify a vector of ones}

\item{a}{square numeric dense matrix of doubles containing the coefficients of the linear system}

\item{b}{numeric vector of doubles giving the right-hand side of the linear system}

\item{maxit}{integer, maximum number of iterations for least squares sequential coordinate descent}

\item{tol}{double, tolerance for convergence of the least squares sequential coordinate descent solver}

\item{min}{double minimum permitted value, use zero for NMF}

\item{max}{double, maximum permitted value, use a large value to avoid an upper bound constraint}

\item{values}{numeric array, either a vector of c(0, 0) to ignore multinomial constraints, or a vector of permitted values in the projection. Must be specified in ascending order and be of at least length 2 and strictly in the range [min, max].}

\item{L0}{the cardinality or L0 truncation to impose on x. This will be incrementally enforced.}

\item{fixed}{a vector of 0 or 1 corresponding to "x" indicating if values in the supplied "x" vector should be fixed (i.e. not updated). To fix, specify 1, otherwise specify 0. If not fixing any values, specify a vector of zeros of length x.}
}
\value{
x vector of least squares solution
}
\description{
Solve \code{a \%*\% x = b} using sequential coordinate descent bounded least squares, where \code{a} is a square matrix and \code{x} and \code{b} are vectors and \code{x} is constrained to values in a specified multinomial distribution and the cardinality of x is truncated to specified L0 value.
}
\details{
See \code{cls} R function documentation for a more convenient interface and details. This function should only be used when speed is paramount. Fields are not nullable. \code{multinomial_bls} first fits a simple bounded model, then takes the largest value and fixes it to the closest value in the multinomial distribution. This process is repeated until until all values have been constrained. L0 regularization is applied to the \code{multinomial_bls} fit by incrementally imposing the penalty and refitting the model with each step into the full truncation.
}
