% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_LossSample}
\alias{Rcpp_LossSample}
\title{Loss of a factorization for each sample}
\usage{
Rcpp_LossSample(w, d, h, A, loss_type, threads)
}
\arguments{
\item{w}{matrix of features x factors (rows x columns)}

\item{d}{vector giving scaling diagonal between w and h, if no scaling specify a vector of 1s of the same length as the number of factors (i.e. number of columns in "w")}

\item{h}{matrix of factors x samples (rows x columns)}

\item{A}{dense matrix of features x samples (rows x columns)}

\item{loss_type}{integer, 1 for "mse" or any other value for "mae"}

\item{threads}{number of threads for OpenMP parallelization, set to 0 to let OpenMP decide and use all available threads}
}
\value{
loss of the factorization for each sample (as mse or mae)
}
\description{
Calculates the loss of "A - wdh" for each sample, where "A" is dense.
}
\details{
This function is for cases where "A" is dense. Use the R function "errorSample" for a more accessible interface and details. This is the Rcpp interface for the fastest possible implementation.
}
