% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_SpLoss}
\alias{Rcpp_SpLoss}
\title{Loss of a factorization}
\usage{
Rcpp_SpLoss(
  w,
  d,
  h,
  A,
  A_weights,
  w_L1,
  w_L2,
  w_PE,
  h_L1,
  h_L2,
  h_PE,
  loss_type
)
}
\arguments{
\item{w}{matrix of features x factors (rows x columns)}

\item{d}{vector giving scaling diagonal between w and h, if no scaling specify a vector of 1s of the same length as the number of factors (i.e. number of columns in "w")}

\item{h}{matrix of factors x samples (rows x columns)}

\item{A}{sparse matrix of features x samples (rows x columns)}

\item{A_weights}{sparse weighting matrix for the loss, the function will not apply any weighting matrix that is not of the same dimensions as "A". For instance, if no weights are to be applied, simply specify a dense matrix of 1 with 1 row and column (of course, in sparse format).}

\item{w_L1}{L1/LASSO regularization on w}

\item{w_L2}{L2/Ridge regression on w}

\item{w_PE}{PE/Pattern Extraction regularization on w}

\item{h_L1}{L1/LASSO regularization on h}

\item{h_L2}{L2/Ridge regression on h}

\item{h_PE}{PE/Pattern Extraction regularization on h}

\item{loss_type}{integer, 1 for "mse" or any other value for "mae"}

\item{threads}{number of threads for OpenMP parallelization, set to 0 to let OpenMP decide and use all available threads}
}
\value{
loss of the factorization (as mse or mae)
}
\description{
Calculates the loss of "A - wdh" with optional weights and one-sided L1, L2, and PE regualrizations, parallelized across columns of "A", where "A" is sparse.
}
\details{
This function is for cases where "A" is a sparse matrix. Use the R function "error" for a more accessible interface and details. This is the Rcpp interface for the fastest possible implementation.
}
