% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_SpSolve}
\alias{Rcpp_SpSolve}
\title{Solve <SpMat>A = WH for H}
\usage{
Rcpp_SpSolve(h, wt, A, threads, L0, L1, L2, PE, values, maxit, tol, min, max)
}
\arguments{
\item{h}{initial sample embeddings model (as factors x samples)}

\item{wt}{transposed feature model dense matrix (as factors x features, note transposition)}

\item{A}{sparse matrix of features x samples}

\item{threads}{integer, threads number of threads to use, or 0 to use all threads as determined by OpenMP}

\item{L0}{integer, L0 sample truncation}

\item{L1}{double, L1/Lasso regularization}

\item{L2}{double, L2/ridge regression}

\item{PE}{double, pattern extraction regularization}

\item{values}{numeric array, either a vector of c(0, 0) to ignore multinomial constraints, or a vector of permitted values in the projection}

\item{maxit}{integer, number of iterations permitted for the sequential coordinate descent least squares solver}

\item{tol}{double, tolerance for convergence of sequential coordinate descent least squares solver}

\item{min}{double minimum permitted value in model (usually 0)}

\item{max}{double maximum permitted value in model (set to very large value to avoid an upper bound)}
}
\value{
h matrix of sample embeddings
}
\description{
See "tlearn" R function documentation for practical details. This function requires full specification as described below. Fields are not nullable.
}
